+++ packages/aws-lambda/lambdas/bin/create-zips.sh
@@ -18,6 +18,8 @@ rm -rf instana-aws-lambda*.tgz
 # Note: ${BUILD_LAMBDAS_WITH-xyz} -> default expansion with default xyz, to avoid bash error "unbound variable"
 # when set -u is active.
 
+echo Building all local tar.gz files.
+
 if [[ -z "${BUILD_LAMBDAS_WITH-}" ]]; then
   echo "Environmant variable BUILD_LAMBDAS_WITH has not been provided, assuming \"npm\" (build with latest npm package)."
 fi
@@ -43,11 +45,11 @@ mv instana-aws-lambda-auto-wrap*.tgz ../aws-lambda/instana-aws-lambda-auto-wrap.
 cd ../aws-lambda
 
 if [[ "${BUILD_LAMBDAS_WITH-npm}" == "npm" ]]; then
-  echo "Building all Lambda zips with the latest npm packages."
+  echo "Building Lambda zip file(s) with the latest npm packages."
 elif [[ "${BUILD_LAMBDAS_WITH-}" == "local" ]]; then
-  echo "Building all Lambda zips with the local tar.gz files."
+  echo "Building Lambda zip file(s) with the local tar.gz files."
 elif [[ "$BUILD_LAMBDAS_WITH" == "layer" ]]; then
-  echo "Building all Lambda zips without @instana/aws-lambda, assuming the AWS Lambda layer \"instana\" is (or will be) configured."
+  echo "Building Lambda zip file(s) without @instana/aws-lambda, assuming the AWS Lambda layer \"instana-nodejs\" is (or will be) configured."
 else
   echo "Unknown option for BUILD_LAMBDAS_WITH: $BUILD_LAMBDAS_WITH"
   echo Aborting.
@@ -56,12 +58,23 @@ fi
 
 popd > /dev/null
 
-for lambda_directory in */ ; do
+if [[ -z $1 ]]; then
+  echo Creating *all* zip files.
+  echo
+  for lambda_directory in */ ; do
+    if [[ -d "$lambda_directory" && ! -L "$lambda_directory" && -e "$lambda_directory/bin/create-zip.sh" ]]; then
+      echo "next directory: $lambda_directory"
+      $lambda_directory/bin/create-zip.sh
+    else
+      echo "skipping directory: $lambda_directory"
+    fi
+  done
+else
+  echo "Creating *only* $1.zip"
+  lambda_directory=$1
   if [[ -d "$lambda_directory" && ! -L "$lambda_directory" && -e "$lambda_directory/bin/create-zip.sh" ]]; then
-    echo "next directory: $lambda_directory"
     $lambda_directory/bin/create-zip.sh
   else
-    echo "skipping directory: $lambda_directory"
+    echo "Cannot create zip file for $lambda_directory, either the directory does not exist or is a symlink or it has no bin/create-zip.sh script."
   fi
-done
-
+fi