+++ packages/aws-lambda/lambdas/bin/rebuild-redeploy-single.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+set -eEuo pipefail
+
+cd `dirname $BASH_SOURCE`/..
+
+bin/create-zips.sh $1
+bin/deploy-demo.sh $1
+